test(cron): consolidate into 4 files, drop legacy cronTasks.ts + cronTasksConcurrency.test.ts
Final layout of cron test coverage:
- cronTasks.scheduling.ts: interval/CRON/duration semantics (19 tests)
- cronTasks.behavior.ts:   public API contract, real time (31 tests)
- cronTasks.parallel.ts:   cronTaskConcurrency > 1 path (8 tests)
- cronTasks.races.regression.ts: impl-aware invariants kept from the
  legacy suite (15 tests) - idle-DB, index presence, query plan,
  projections, lock prolong / mid-prolong race, revert-on-cancel race,
  DB fault tolerance (5 scenarios). Each test has a JSDoc banner
  describing the invariant it pins.

Delete:
- test/cronTasks.ts: superseded by the 4 files above.
- test/cronTasksConcurrency.test.ts: absorbed into cronTasks.parallel.ts
  (lazy-start case merged; overlap/serialisation/stop cases already
  covered with broader assertions).

Net: 83 -> 73 tests, but scenario coverage is unchanged (it.each on
flaky `times(10)` / `times(3)` variants collapsed to single or small
deterministic runs). Parallel path that was previously only covered in
cronTasksConcurrency.test.ts is now first-class in cronTasks.parallel.ts
alongside serial sanity-check and pendingWake/runnerStopPromise regressions.
